+++ source/Examples/Resources/IEEnhancedSecurityConfiguration/2-IEEnhancedSecurityConfiguration_DisableForUsers_Config.ps1
@@ -20,7 +20,7 @@
 <#
     .DESCRIPTION
         This configuration will disable the IE Enhanced Security Configuration for
-        administrators.
+        users.
 #>
 Configuration IEEnhancedSecurityConfiguration_DisableForUsers_Config
 {